% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_toy_communities.R
\name{assign_individual_mass}
\alias{assign_individual_mass}
\title{Draw a mass for an individual}
\usage{
assign_individual_mass(species_id, bsd, assumptions = list(sd = 0.1))
}
\arguments{
\item{species_id}{int species identifier}

\item{bsd}{vector of mean masses}

\item{assumptions}{list including sd}
}
\value{
mass
}
\description{
Draw a mass for an individual
}
