% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_toy_communities.R
\name{assign_mean_mass}
\alias{assign_mean_mass}
\title{Assign mean masses to toy species}
\usage{
assign_mean_mass(S, min_size, max_size, assumption = "log_uniform")
}
\arguments{
\item{S}{nb species}

\item{min_size}{in g, linear}

\item{max_size}{in g, linear}

\item{assumption}{"log_uniform", "hutchinson", ...}
}
\value{
vector of mean body sizes
}
\description{
Assign mean masses to toy species
}
