% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_toy_communities.R
\name{assign_species_id}
\alias{assign_species_id}
\title{Assign individuals to species}
\usage{
assign_species_id(S, N, assumption = "uniform")
}
\arguments{
\item{S}{nb species}

\item{N}{nb individuals}

\item{assumption}{"uniform", ...}
}
\value{
vector of species assignments
}
\description{
Assign individuals to species
}
