% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{fit_power_law}
\alias{fit_power_law}
\title{Fit powerlaw to community ISD}
\usage{
fit_power_law(dat, estimate_min = TRUE)
}
\arguments{
\item{dat}{community df with `individual_sizes` column}

\item{estimate_min}{whether to estimate the minimum for the power law, or use the minimum of `individual_sizes`. Defaults TRUE for estimate.}
}
\value{
dataframe of simulated dat drawn from fitted power law
}
\description{
Fit powerlaw to community ISD
}
