% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toy_portal_data.R
\name{get_toy_portal_data}
\alias{get_toy_portal_data}
\title{Get just a little bit of Portal data}
\usage{
get_toy_portal_data()
}
\value{
a df of Portal control plots from 1994-1995
}
\description{
Get just a little bit of Portal data
}
