% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gmm_pdf}
\alias{plot_gmm_pdf}
\title{Plot PDF from GMM}
\usage{
plot_gmm_pdf(gmm_pdf, pdf_name = NULL)
}
\arguments{
\item{gmm_pdf}{the pdf}

\item{pdf_name}{the name of the community, defaults null}
}
\value{
density plot
}
\description{
Plot PDF from GMM
}
