% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ied}
\alias{plot_ied}
\title{Plot ied}
\usage{
plot_ied(ied, ied_name = NULL)
}
\arguments{
\item{ied}{result of `make_isd`}

\item{ied_name}{community name, defaults NULL}
}
\value{
plot of IED
}
\description{
Plot ied
}
