% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{process_neon_data}
\alias{process_neon_data}
\title{Process NEON data
Load NEON mammal trapping data
Filter to target taxa and remove problem sites following Read et al 2018 Ecography: https://figshare.com/articles/A_thermal_gradient_of_trait_similarity_across_North_America/5339407}
\usage{
process_neon_data()
}
\value{
TRUE if successful
}
\description{
Process NEON data
Load NEON mammal trapping data
Filter to target taxa and remove problem sites following Read et al 2018 Ecography: https://figshare.com/articles/A_thermal_gradient_of_trait_similarity_across_North_America/5339407
}
