% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{sample_pl}
\alias{sample_pl}
\title{Sample from fitted powerlaw}
\usage{
sample_pl(dat, plpars, nsamples)
}
\arguments{
\item{dat}{community dataframe with individual_sizes column}

\item{plpars}{result of get_pl_pars(dat)}

\item{nsamples}{how many to draw}
}
\value{
matrix of samples
}
\description{
Sample from fitted powerlaw
}
