% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{loglik_pl}
\alias{loglik_pl}
\title{Get loglik of empirical from powerlaw}
\usage{
loglik_pl(dat, plpars)
}
\arguments{
\item{dat}{community dataframe with individual_sizes column}

\item{plpars}{result of get_pl_pars(dat)}
}
\value{
loglikelihood
}
\description{
Get loglik of empirical from powerlaw
}
